# docs/04-API명세-openapi.yaml
openapi: 3.0.3
info:
  title: HOOPLY API
  version: 1.0.0
  description: 농구 동호회 팀 매칭 서비스. 인증은 Bearer JWT(access 30분/refresh 14일). 모든 4xx/5xx 본문은 ErrorResponse {code, message, details[]}.
servers:
- url: https://hooply-backend.onrender.com/api/v1
  description: 운영
- url: http://localhost:8000/api/v1
  description: 로컬
security:
- bearerAuth: []
paths:
  /auth/signup:
    post:
      summary: 이메일 회원가입
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      security: []
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '409':
          description: EMAIL_DUPLICATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: 이메일 로그인
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      security: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: INVALID_CREDENTIALS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/kakao/login-url:
    get:
      summary: 카카오 인가 URL 생성
      tags:
      - 인증 · 프로필
      parameters:
      - name: redirect_uri
        in: query
        required: false
        schema:
          type: string
          nullable: true
      security: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KakaoLoginUrl'
        '401':
          description: KAKAO_AUTH_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/kakao/callback:
    get:
      summary: 카카오 콜백 (가입/로그인)
      tags:
      - 인증 · 프로필
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: state
        in: query
        required: true
        schema:
          type: string
      - name: redirect_uri
        in: query
        required: false
        schema:
          type: string
          nullable: true
      security: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KakaoTokenPair'
        '401':
          description: KAKAO_AUTH_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/kakao/link:
    post:
      summary: 기존 계정에 카카오 연결
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KakaoLinkRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '401':
          description: KAKAO_AUTH_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: IDENTITY_ALREADY_LINKED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: 토큰 갱신
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: TOKEN_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: 로그아웃 (refresh 토큰 폐기)
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      security: []
      responses:
        '204':
          description: 성공
  /auth/password/forgot:
    post:
      summary: 비밀번호 재설정 요청
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      security: []
      responses:
        '202':
          description: 성공
  /auth/password/reset:
    post:
      summary: 비밀번호 재설정
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      security: []
      responses:
        '200':
          description: 성공
        '400':
          description: TOKEN_INVALID_OR_EXPIRED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me:
    get:
      summary: 내 정보 조회
      tags:
      - 인증 · 프로필
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
    delete:
      summary: 계정 삭제
      tags:
      - 인증 · 프로필
      responses:
        '204':
          description: 성공
        '422':
          description: CANNOT_DEMOTE_LAST_MANAGER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: 내 프로필 수정
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
  /me/password:
    post:
      summary: 비밀번호 변경
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: INVALID_CREDENTIALS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /me/avatar:
    post:
      summary: 프로필 사진 등록 · 교체
      tags:
      - 인증 · 프로필
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvatarIn'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
        '400':
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: 프로필 사진 삭제
      tags:
      - 인증 · 프로필
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetail'
  /users/{user_id}/avatar:
    get:
      summary: 프로필 사진 이미지
      tags:
      - 인증 · 프로필
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: v
        in: query
        required: false
        schema:
          type: string
          nullable: true
      responses:
        '200':
          description: 성공
        '404':
          description: Not Found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AvatarIn:
      properties:
        data_url:
          type: string
          maxLength: 700000
      type: object
      required:
      - data_url
...
    ChangePasswordRequest:
      properties:
        current_password:
          type: string
          format: password
          writeOnly: true
        new_password:
          type: string
          maxLength: 72
          minLength: 8
          format: password
          writeOnly: true
      type: object
      required:
      - current_password
      - new_password
...
    ErrorDetail:
      properties:
        field:
          type: string
          nullable: true
        reason:
          type: string
        context:
          additionalProperties: true
          type: object
          nullable: true
      type: object
      required:
      - reason
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          type: array
      type: object
      required:
      - code
      - message
...
    ForgotPasswordRequest:
      properties:
        email:
          type: string
          format: email
      type: object
      required:
      - email
...
    IdentityView:
      properties:
        provider:
          type: string
        linked_at: {}
      type: object
      required:
      - provider
      - linked_at
...
    KakaoLinkRequest:
      properties:
        code:
          type: string
        state:
          type: string
        redirect_uri:
          type: string
          nullable: true
      type: object
      required:
      - code
      - state
    KakaoLoginUrl:
      properties:
        url:
          type: string
        state:
          type: string
      type: object
      required:
      - url
      - state
    KakaoTokenPair:
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        is_new:
          type: boolean
      type: object
      required:
      - access_token
      - refresh_token
      - is_new
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
      type: object
      required:
      - email
      - password
    LogoutRequest:
      properties:
        refresh_token:
          type: string
      type: object
      required:
      - refresh_token
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
      type: object
      required:
      - refresh_token
    ResetPasswordRequest:
      properties:
        token:
          type: string
        new_password:
          type: string
          maxLength: 72
          minLength: 8
          format: password
          writeOnly: true
      type: object
      required:
      - token
      - new_password
...
    SignupRequest:
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          maxLength: 72
          minLength: 8
          format: password
          writeOnly: true
        name:
          type: string
          maxLength: 50
          minLength: 1
        nickname:
          type: string
          maxLength: 50
          nullable: true
        height_cm:
          type: integer
          maximum: 250.0
          minimum: 120.0
          nullable: true
      type: object
      required:
      - email
      - password
      - name
...
    TokenPair:
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
      type: object
      required:
      - access_token
      - refresh_token
...
    UserDetail:
      properties:
        id:
          type: integer
        email:
          type: string
          nullable: true
        name:
          type: string
        nickname:
          type: string
          nullable: true
        profile_image_url:
          type: string
          nullable: true
        height_cm:
          type: integer
          nullable: true
        global_role:
          type: string
        onboarding_completed:
          type: boolean
        primary_team_id:
          type: integer
          nullable: true
        identities:
          items:
            $ref: '#/components/schemas/IdentityView'
          type: array
        tutorial_state:
          type: string
        tutorial_path:
          type: string
          nullable: true
        tutorial_tips_seen:
          items:
            type: string
          type: array
      type: object
      required:
      - id
      - email
      - name
      - nickname
      - profile_image_url
      - height_cm
      - global_role
      - onboarding_completed
...
    UserUpdate:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          nullable: true
        nickname:
          type: string
          maxLength: 50
          nullable: true
        profile_image_url:
          type: string
          maxLength: 500
          nullable: true
        height_cm:
          type: integer
          maximum: 250.0
          minimum: 120.0
          nullable: true
        primary_team_id:
          type: integer
          nullable: true
      type: object
...